refactor: drop auth entry bundle
