--- storefronts/features/auth/sdk-auth-entry.js
@@ -1,32 +1,0 @@
-import auth from './index.js';
-import { init, loadConfig } from './init.js';
-import {
-  lookupRedirectUrl,
-  lookupDashboardHomeUrl
-} from '../../../supabase/authHelpers.js';
-import * as currency from '../currency/index.js';
-import { getConfig } from '../config/globalConfig.js';
-
-if (typeof globalThis.setSelectedCurrency !== 'function') {
-  globalThis.setSelectedCurrency = () => {};
-}
-
-const storeRedirects = { lookupRedirectUrl, lookupDashboardHomeUrl };
-const SMOOTHR_CONFIG = getConfig();
-
-const Smoothr = { auth, loadConfig, storeRedirects, currency, SMOOTHR_CONFIG };
-export { auth, loadConfig, storeRedirects, currency, SMOOTHR_CONFIG };
-export default Smoothr;
-
-if (typeof window !== 'undefined') {
-  // ✅ Assign modern global
-  window.Smoothr = { ...(window.Smoothr || {}), ...Smoothr };
-
-  // ✅ TEMP PATCH for legacy build check (scripts/check-sdk.js)
-  // This ensures string "window.smoothr" is present in the output
-  window.smoothr = window.smoothr || {};
-  window.smoothr.auth = window.Smoothr?.auth;
-}
-
-// ✅ Initialize for backward compatibility (legacy clients)
-init();
