+++ storefronts/tests/sdk/load-config-merge.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import * as core from '../../features/auth/sdk-auth-entry.js';
+import * as core from '../../features/auth/init.js';
 
 vi.mock('../../features/auth/index.js', () => {
   const authMock = {
@@ -92,7 +92,7 @@ describe('loadConfig merge', () => {
   it('preserves existing config values', async () => {
     console.log('Starting test: loadConfig merge');
     const { loadConfig } = await import(
-      '../../features/auth/sdk-auth-entry.js'
+      '../../features/auth/init.js'
     );
     await loadConfig('00000000-0000-0000-0000-000000000000');
